+++ changelog.md
@@ -1,5 +1,9 @@
 
+
+
 *	2020.4.12
+
+	+	imgPicker.js: image types added.
 
 	+	imgPicker.js: listenPage added.